Reworked inventory and action system
- Weapons can be equipped now and bind to the action slot automatically
- Proper support for one-handed and two-handed weapons
- Damage is based on weapon now
- TODO: Clean up weapon stats
- TODO: Finalize concept of weapons vs. skills vs. consuming items in action slots
- TODO: Clean up action slots UI flow and dependencies

+++ common/interfaces/IDamageable.cs
@@ -5,5 +5,5 @@ public interface IDamageable
 	/// <summary>
 	/// Called when the object takes damage
 	/// </summary>
-	void TakeDamage(int amount, Vector3 attackDirection);
+	void TakeDamage(float amount, Vector3 attackDirection);
 }
